MagicScaling to Scaling
