+++ OneMore/OneMore.csproj
@@ -694,7 +694,6 @@
     <Compile Include="Helpers\HttpClientFactory.cs" />
     <Compile Include="Helpers\ImageDetector.cs" />
     <Compile Include="Helpers\Loggable.cs" />
-    <Compile Include="Helpers\MagicScaling.cs" />
     <Compile Include="Helpers\Office\Excel.cs" />
     <Compile Include="Helpers\Office\Models\OutlookEnums.cs" />
     <Compile Include="Helpers\Office\Models\OutlookTaskFolders.cs" />
